fix(cli): move overlay fixture tests out of src/ so generated CLIs build
PR #16056 removed the dev fixture (cli/openapi-fixture/**) from
shipped CLI output but left two #[test] blocks in
generators/cli/sdk/src/openapi/overlay.rs that
include_str!("../../cli/openapi-fixture/openapi.yaml"). Because src/
files ship verbatim and SDK_IGNORE prunes the fixture path, every
generated CLI was failing `cargo build --locked --all-features --tests`
at overlay.rs:1835 and :1894 with `couldn't read … openapi.yaml`.

The two tests were pure integration coverage of the dev fixture's
full spec, not unit tests of overlay logic — they belong in
tests/. Moving them to a new tests/overlay_fixture.rs (added to
SDK_IGNORE) keeps the dev-tree coverage and stops shipping the
broken include_str! sites to users. The CliApp::build_doc() test is
rewritten to use the already-public load_openapi_spec pipeline so we
don't have to widen any pub(crate) symbols.

Adds a regression-guard TS test
(generators/cli/src/__test__/sdkTemplateHygiene.test.ts) that scans
every .rs under sdk/src/ and fails if any include_str!/include_bytes!/
include! escapes the src/ tree via ../../ — catches the broader bug
class, not just this specific path.
